+++ src/components/EthAddressAvatar/declarations.d.ts
